Migrate build_method tests to Node Namespace
The build method test have been migrated to the new Namespace. Not
all the tests are currently running as they require the self node to
be supported.

As these are the last tests to migrate, the old Node spec has been
deleted
